samples/pygithub/requirements.txt
black
termcolor
PyGithub